requirements.txt
streamlit==1.31.0
chromadb==0.4.22
pdfplumber==0.10.3
PyPDF2==3.0.1
langchain==0.1.6
langchain-community==0.0.20
sentence-transformers==2.3.1
google-generativeai==0.3.2
python-dotenv==1.0.1
